+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2023-05-04
+## [2.7.0] - 2023-05-08
 
 ### Changed
 
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Remove SharePoint Server 2013 from supported versions
 - Remove azure-pipelines.yml and SPSWakeUp.psd1 files
 - SPSWakeUP.ps1
- - BREAKING CHANGE: Remove InputFile switch parameter and update examples
+ - BREAKING CHANGE: Remove InputFile and OnlyRootWeb switch parameters and update examples
  - Add Transcript boolean parameter and update examples
  - BREAKING CHANGE: Remove logging functions and logging variables
  - Replace Write-LogContent function with Write-Output CmdLet
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Get-SPSHSNCUrl
  - Update Disable-LoopbackCheck, Add-HostsEntry, Get-SPSSitesUrl and Get-SPSWebAppUrl functions
  - Remove unnecessary comments
+ - Add Get-SPSAdminUrl function to improve central admin urls warmup
 
 ## [2.6.1] - 2023-01-18